Fixes for picard_solver and anderson_solver (wrong convergence)
Both solvers would converge, but the solution (uvel,vvel)
was not qualitatively the same as the EVP solution; velocities were
too small.

For picard_solver, it was due to the fact that the initial residual stored
and used to check for convergence was not the initial residual at the start of the nonlinear iteration,
but the last residual at the end of the first FGMRES iteration. Since
this residual was smaller, convergence would be "faster" but the solution was not the same;
it was not converged enough in the sense of the nonlinear residual. This was an implementation error when
the nonlinear convergence logic was removed from the fgmres solver.

For anderson_solver, it was due to using the fixed point residual norm to check for convergence.
Similarly, this residual is smaller than the nonlinear residual and thus leads to a "faster" convergence,
but to a different solution (not converged enough in the sense of the nonlinear residual).