matvec: change Au,Av from intent(out) to intent(inout)
According to The Fortran 2003 Handbook [1], the Fortran standard states that
arguments with intent(out) are left uninitialized if the subroutine/function
does not set them. This is dangerous here since the Au, Av arguments are set
in matvec only for grid cells where there is ice, so any grid cells with no
ice that were previously initialized (eg. to zero) in the actual arguments
could be left with garbage values after the call to matvec (this does not
happen with the Intel compiler but it's still better to follow the standard).

[1] J. C. Adams, W. S. Brainerd, R. A. Hendrickson, R. E. Maine, J. T. Martin,
    and B. T. Smith, The Fortran 2003 Handbook. London: Springer London, 2009.