WIP: TOREMOVE: Add conda environment to run on personal computers (#393)
* machines: add env and Macro files for conda on Linux and macOS

* machines: add environment.yml conda specification

* doc: add section on laptop computers

Uses the conda environment.

* doc: recommend not to touch the 'cice' conda environment

* machines: add CFLAGS_HOST to conda_macos for recent macOS

Recent versions of Xcode and macOS do not install the system C header files
to /usr/include, but keep them in the MacOSX SDK.
Thus we need to explicitely pass their location to the compiler.

* Makefile: remove uneeded (and faulty) logic around CFLAGS_HOST

The CFLAG_HOST variable was added in CICE-Consortium/CICE#307, and
some Makefile logic was added to define the variable to be blank if
the included Macros file does not define it.

However, the Make syntax is wrong, as 'ifndef' takes a variable name and not
a reference to a variable [1], so

    ifndef $(CFLAGS_HOST)

should have been written

    ifndef CFLAGS_HOST

The effect of this error is to invert the logic of the check (!) since if
CFLAGS_HOST is defined, Make will check if a variable with name equal to
whatever CFLAGS_HOST is defined to be exists, which will most probably be false,
and the conditional will then evaluate to true and CFLAG_HOSTS will be redefined
to be blank, defeating the whole purpose of the flag.

Since there's no harm in Make referencing and empty variable, fix this by just
removing the conditional check.

[1] https://www.gnu.org/software/make/manual/html_node/Conditional-Syntax.html

* doc: use curl instead of wget on macOS

wget is not installed by default

* machines: conda: invoke 'conda info --base' using $CONDA_EXE

This variable is always defined to the path to the conda executable
when the login shell is properly initialized.

Calling `conda info --base` works if the shell initialization procedure
puts the conda executable in the $PATH, which is the case for Bash and Zsh
but not for tcsh.

* doc: conda: add initialization instructions for alternative login shells

The Miniconda installer only initializes Bash on macOS and Linux,
so additional steps needs to be taken if your login shell is different.

Add instructions for tcsh, zsh, fish and xonsh.

* machines: conda: add error checking to env files

Let's check if the conda executable is found and if the cice
conda environment exists.

* doc: separate Miniconda installation and conda initialization

- Move the documentation for the conda configuration under "Porting"
- Move the instructions for initializing the user's shell for use with conda
  to a speparate section
- Add a new section for manually initializing the user's shell if they do not
  want to modify their startup files
- Emphasize which steps are "first-time setup" steps
- Add details about how the conda environment is activated during build/run

* doc: conda: mention symlinks and creating a complete port