WIP: TOREMOVE: scripts: add support for setting `ICE_CPPDEFS` in set_env.* options
Allow using the present case/test configuration mechanism
(set_env.* in configuration/scripts/options/) to add preprocessor macros.

cice.setup already allows for modifying cice.setttings using set_env.*
options, so the only change needed is to add a default value for
`ICE_CPPDEFS` in the default cice.settings file, and not resetting
this variable in cice.build.

Note that contrary to what was proposed in CICE-Consortium/CICE#445, we
do not implement a similar ICE_SLIBS variable, that could be used to add
linker flags (such as optional libraries), because the way in which
libraries are made accessible and linked with an executable vary
depending on the machine in use, so such modifications are best made in
the Macros.<machine>_<compiler> and env.<machine>_<compiler> files.

Closes CICE-Consortium/CICE#445