ice_dyn_vp: initialize 'stPr' and 'zetaD' to zero
Since 'stPr' and 'zetaD' are initialized with a loop over icellt, but
are used on loops over icellu, it is possible that these latter loops
use uninitialized values since iceumask and icetmask can differ at some
grid points since they are not defined using the same criteria (see
ice_dyn_shared::dyn_prep1 and ice_dyn_shared::dyn_prep2).

To avoid using unitialized values, we initialize the 'stPr' and 'zetaD'
arrays to zero, so that any index corresponding to a cell with no ice
is set to zero.

This mimics what is done for EVP, where the 'str' array is initialized
to zero in ice_dyn_evp::stress.