ice_grid: initialize 'tarea' to 0
In subroutine 'init_grid2', the array 'tarea' is initialized in a loop on
indices {ilo, ihi} and {jlo, jhi}, but is then used in subroutine 'makemask' in
a loop on indices {1, nx_block} and {1, ny_block}, causing an uninitialized
value to be used (at line 1693).

Initialize the whole array to zero first, to avoid using uninitialized values.
The initialization to zero is actually commented out, following
https://github.com/CICE-Consortium/CICE/pull/180, so uncomment it.