ice_flux: remove CICE_IN_NEMO CPP around sst initialization
WIP IMPROVE MESSAGE

'cice_init' is called a second time in sbcice_cice, so the
value set here does not matter, and if we compile with `-init=snan,arrays` and
`-DCICE_IN_NEMO`, not providing a value makes the code abort in
'ice_init::set_state_var' when 'sst' is used in a condition.

As such, remove the '#ifdef CICE_IN_NEMO' around the initialization of the
'sst' array to 'Tf', so that it is always initialized.

(cherry picked from commit 06de41149e1b24a7d64bebb07221d47d447ec6be)